+++ model.py
@@ -35,7 +35,8 @@ class Generator(nn.Module):
 
         #label_embedding = self.embedding(label)
 
-        x = torch.cat((noise,label),dim=1)
+        x = torch.cat((noise,label.float()),dim=1)
+        #print(label.type)
         x = x.view(-1,110,1,1)
 
         x = self.layer1(x)
